+++ packages/redux-dynamic-modules-core/.npmignore
