+++ packages/redux-dynamic-modules-core/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "redux-dynamic-modules-react",
+    "name": "redux-dynamic-modules-core",
     "version": "3.3.1",
     "description": "Modularize the redux app by dynamically loading reducers, state and sagas",
     "repository": {
@@ -12,7 +12,8 @@
         "prepublish": "npm run build:prod && npm run test",
         "build": "npm run clean && tsc && webpack --mode development --display-modules --progress --display-error-details",
         "build:prod": "npm run clean && tsc && webpack --mode development --display-error-details && webpack --mode production --display-modules --progress --display-error-details",
-        "clean": "rimraf lib/ dist/"
+        "clean": "rimraf lib/ dist/",
+        "test": "jest"
     },
     "keywords": [
         "react",
@@ -29,8 +30,6 @@
         "@types/react-redux": "6.0.11",
         "@types/redux": "3.6.0",
         "jest": "^23.5.0",
-        "react": "16.8.0",
-        "react-redux": "6.0.0",
         "redux": "4.0.1",
         "rimraf": "^2.6.2",
         "ts-jest": "^23.1.4",
@@ -42,12 +41,29 @@
         "webpack-bundle-analyzer": "^2.13.1",
         "webpack-cli": "^3.1.0"
     },
+    "jest": {
+        "transform": {
+            "^.+\\.tsx?$": "ts-jest"
+        },
+        "verbose": false,
+        "testMatch": [
+            "**/src/__tests__/**/(*.)+(spec|test).ts?(x)"
+        ],
+        "moduleFileExtensions": [
+            "ts",
+            "tsx",
+            "js",
+            "jsx",
+            "json",
+            "node"
+        ]
+    },
     "peerDependencies": {
-        "react": ">= 15.0.0",
-        "react-redux": ">= 5.0.0"
+        "redux": ">= 3.0.0"
     },
     "dependencies": {
-        "redux-dynamic-modules": "^3.3.1"
+        "redux-devtools-extension": "^2.13.7",
+        "redux-dynamic-middlewares": "^1.0.0"
     },
     "gitHead": "56fcc21ecad212dae465defe7239c7b08a137e12"
 }
