+++ packages/redux-dynamic-modules-core/src/Utils/Flatten.ts
