+++ packages/redux-dynamic-modules-core/src/Utils/RefCounter.ts
