+++ packages/redux-dynamic-modules-core/src/__tests__/Managers/ModuleManager.test.ts
