+++ packages/redux-dynamic-modules-core/src/__tests__/Utils/Flatten.test.ts
