+++ packages/redux-dynamic-modules-core/src/index.ts
@@ -1,0 +1,7 @@
+export * from "./Contracts";
+export * from "./ModuleStore";
+export * from "./Utils/ComparableMap";
+export * from "./Utils/RefCounter";
+export * from "./Managers/MiddlewareManager";
+export * from "./Managers/RefCountedManager";
+//Dummy change to keep lerna happy
