+++ packages/redux-dynamic-modules-core/tslint.json
