+++ packages/redux-dynamic-modules-core/webpack.config.js
