--- packages/redux-dynamic-modules-react/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./DynamicModuleHook";
-export * from "./DynamicModuleLoader";
