+++ packages/redux-dynamic-modules/src/index.ts
@@ -1,7 +1,3 @@
-export * from "./Contracts";
-export * from "./ModuleStore";
-export * from "./Utils/ComparableMap";
-export * from "./Utils/RefCounter";
-export * from "./Managers/MiddlewareManager";
-export * from "./Managers/RefCountedManager";
-//Dummy change to keep lerna happy
+export * from "redux-dynamic-modules-core";
+export * from "./DynamicModuleHook";
+export * from "./DynamicModuleLoader";
