+++ packages/redux-dynamic-modules/tsconfig.json
@@ -16,7 +16,8 @@
         "outDir": "lib"
     },
     "include": [
-        "src/**/*.ts"
+        "src/**/*.ts",
+        "src/**/*.tsx"
     ],
     "exclude": [
         "src/__tests__/**/*.ts"
