+++ packages/typescript-example/src/Components/Order.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { connect } from "react-redux";
-import { DynamicModuleLoader } from "redux-dynamic-modules-react";
+import { DynamicModuleLoader } from "redux-dynamic-modules";
 import { getOrderModules } from "../OrderModule/OrderModule";
 import { IOrderAwareState } from "../OrderModule/OrderContracts";
 import { getUserPreferences } from "../SettingsModule/SettingsSelectors";
