+++ packages/widgets-example/src/widgets/hacker-news/index.js
@@ -1,6 +1,6 @@
 import { ConnectedHackerNews } from "./component/hacker-news-component";
 import { getHackerNewsModule } from "./redux/hacker-news-module";
-import { DynamicModuleLoader } from "redux-dynamic-modules-react";
+import { DynamicModuleLoader } from "redux-dynamic-modules";
 import * as React from "react";
 
 export default function DynamicHackerNews() {
