+++ packages/widgets-example/src/widgets/weather/index.js
@@ -1,6 +1,6 @@
 import { ConnectedWeather } from "./component/weather-component";
 import { getWeatherModule } from "./redux/weather-module";
-import { DynamicModuleLoader } from "redux-dynamic-modules-react";
+import { DynamicModuleLoader } from "redux-dynamic-modules";
 import * as React from "react";
 
 export default function Dynamic() {
